restructure ssa cache code in separate packages
