+++ internal/controller/object/syncer.go
@@ -2,6 +2,7 @@ package object
 
 import (
 	"context"
+	"github.com/crossplane-contrib/provider-kubernetes/pkg/kube/client/ssa/cache/state"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -14,7 +15,6 @@ import (
 	"github.com/crossplane/crossplane-runtime/pkg/resource"
 
 	"github.com/crossplane-contrib/provider-kubernetes/apis/object/v1alpha2"
-	"github.com/crossplane-contrib/provider-kubernetes/pkg/kube/client/ssa"
 )
 
 // PatchingResourceSyncer is a ResourceSyncer that syncs objects by patching
@@ -66,7 +66,7 @@ func (p *PatchingResourceSyncer) SyncResource(ctx context.Context, obj *v1alpha2
 type SSAResourceSyncer struct {
 	client              client.Client
 	extractor           applymetav1.UnstructuredExtractor
-	desiredStateCacheFn func() ssa.StateCache
+	desiredStateCacheFn func() state.StateCache
 }
 
 // GetObservedState returns the object's observed state by extracting the
