+++ pkg/kube/client/ssa/cache/extractor/caching_unstructured_extractor_test.go
@@ -1,4 +1,4 @@
-package ssa
+package extractor
 
 import (
 	"context"
@@ -222,7 +222,7 @@ func TestExtract(t *testing.T) {
 			}
 
 			cache := &GVKParserCache{
-				store: map[schema.GroupVersion]*GVKParserCacheEntry{},
+				store: map[schema.GroupVersion]*gvkParserCacheEntry{},
 			}
 			ext, err := NewCachingUnstructuredExtractor(context.TODO(), dc, cache)
 			if err != nil {
@@ -652,7 +652,7 @@ func TestParserCaching(t *testing.T) {
 	}
 
 	cache := &GVKParserCache{
-		store: map[schema.GroupVersion]*GVKParserCacheEntry{},
+		store: map[schema.GroupVersion]*gvkParserCacheEntry{},
 	}
 	ext, err := NewCachingUnstructuredExtractor(context.TODO(), dc, cache)
 	if err != nil {
@@ -766,7 +766,7 @@ func TestCachingMultipleExtractors(t *testing.T) {
 	}
 
 	cache := &GVKParserCache{
-		store: map[schema.GroupVersion]*GVKParserCacheEntry{},
+		store: map[schema.GroupVersion]*gvkParserCacheEntry{},
 	}
 	ext, err := NewCachingUnstructuredExtractor(context.TODO(), dc, cache)
 	if err != nil {
@@ -826,7 +826,7 @@ func TestParserCachingEmptyEtag(t *testing.T) {
 	}
 
 	cache := &GVKParserCache{
-		store: map[schema.GroupVersion]*GVKParserCacheEntry{},
+		store: map[schema.GroupVersion]*gvkParserCacheEntry{},
 	}
 	ext, err := NewCachingUnstructuredExtractor(context.TODO(), dc, cache)
 	if err != nil {
