+++ pkg/kube/client/ssa/cache/extractor/openapi_groupversion_test.go
@@ -1,4 +1,4 @@
-package ssa
+package extractor
 
 import (
 	"context"
