+++ pkg/kube/client/ssa/cache/extractor/test/k8s_objects_for_extraction/1_extracted.json
