+++ pkg/kube/client/ssa/cache/extractor/test/k8s_objects_for_extraction/2_for_extraction.json
