+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/apps.v1.json
