+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/core.v1.json
