+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/nop.example.org.v1alpha1.json
