+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/pkg.crossplane.io.v1.json
