+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/ref_validation_test/brokenrefs-multiple-nop.example.org.v1alpha1.json
