+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/ref_validation_test/remoteRef-parentFolder.json
