+++ pkg/kube/client/ssa/cache/extractor/test/openapi_schemas/ref_validation_test/urlRef-another-server-https.json
