+++ pkg/kube/client/ssa/cache/state/state_cache.go
@@ -1,4 +1,4 @@
-package ssa
+package state
 
 import (
 	"crypto/sha256"
